https://www.assembla.com/spaces/erlide/tickets/406-ErlangEditor-getModule-can-return-null--not-all-call-sites-handle-that-case

+++ org.erlide.ui/src/org/erlide/ui/actions/ClearCacheAction.java
@@ -21,7 +21,8 @@ public class ClearCacheAction extends TextEditorAction {
 	private final ErlangEditor erlangEditor;
 	private final String NOPARSE_CACHE_SUFFIX = ".noparse";
 	private final String SCANNER_CACHE_SUFFIX = ".scan";
-	private String suffixes[] = {NOPARSE_CACHE_SUFFIX, SCANNER_CACHE_SUFFIX};
+	private final String suffixes[] = { NOPARSE_CACHE_SUFFIX,
+			SCANNER_CACHE_SUFFIX };
 
 	public ClearCacheAction(final ResourceBundle bundle, final String prefix,
 			final ErlangEditor erlangEditor) {
@@ -37,11 +38,15 @@ public class ClearCacheAction extends TextEditorAction {
 	@Override
 	public void run() {
 		final IErlModule module = erlangEditor.getModule();
+		if (module == null) {
+			return;
+		}
 		for (String suffix : suffixes) {
-			final String cacheFileOSPath = ErlangPlugin.getDefault()
-			.getStateLocation().append(
-					ErlScanner.createScannerModuleName(module)
-					+ suffix).toOSString();
+			final String cacheFileOSPath = ErlangPlugin
+					.getDefault()
+					.getStateLocation()
+					.append(ErlScanner.createScannerModuleName(module) + suffix)
+					.toOSString();
 			final File cacheFile = new File(cacheFileOSPath);
 			cacheFile.delete();
 		}
